refactor(): Move pkg to common
