+++ backend/common/utils/auth.go
@@ -1,14 +1,7 @@
-// Package auth
-// Author: Samoei Oloo
-// Created: 2025-03-28
-// License: None
-//
-// This script is responsible for JWT generation
-package auth
+package utils
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"log"
 	"net/http"
@@ -16,34 +9,28 @@ import (
 	"strings"
 	"time"
 
-	"github.com/samoei-ftw/specno/backend/common/models"
-	repo "github.com/samoei-ftw/specno/backend/common/utils"
-	user_repo "github.com/samoei-ftw/specno/backend/user_service/internal/repo"
-	"github.com/samoei-ftw/specno/backend/user_service/internal/services"
-	"golang.org/x/crypto/bcrypt"
-
 	"github.com/golang-jwt/jwt/v4"
 )
 
 var jwtKey []byte
 
 func init() {
-	// Get JWT secret key
 	secret := os.Getenv("JWT_SECRET")
 	if secret == "" {
 		log.Fatal("JWT_SECRET not set in environment variables")
 	}
 	jwtKey = []byte(secret)
 }
+
 type contextKey string
 
 const ClaimsKey contextKey = "claims"
+
 type Claims struct {
 	UserID int `json:"user_id"`
 	jwt.RegisteredClaims
 }
 
-// Generate a JWT token
 func GenerateToken(userID int) (string, error) {
 	claims := Claims{
 		UserID: userID,
@@ -55,6 +42,7 @@ func GenerateToken(userID int) (string, error) {
 	return token.SignedString(jwtKey)
 }
 
+// returns its claims of parsed jwt string
 func ParseJWT(tokenStr string) (*Claims, error) {
 	token, err := jwt.ParseWithClaims(tokenStr, &Claims{}, func(token *jwt.Token) (interface{}, error) {
 		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
@@ -62,15 +50,12 @@ func ParseJWT(tokenStr string) (*Claims, error) {
 		}
 		return jwtKey, nil
 	})
-
 	if err != nil {
 		return nil, err
 	}
-
 	if claims, ok := token.Claims.(*Claims); ok && token.Valid {
 		return claims, nil
 	}
-
 	return nil, errors.New("invalid token")
 }
 
@@ -79,11 +64,9 @@ func ExtractTokenFromHeader(r *http.Request) (string, error) {
 	if authHeader == "" {
 		return "", errors.New("authorization header missing")
 	}
-
 	if !strings.HasPrefix(authHeader, "Bearer ") {
 		return "", errors.New("invalid authorization header format")
 	}
-
 	return strings.TrimPrefix(authHeader, "Bearer "), nil
 }
 
@@ -92,62 +75,21 @@ func ValidateToken(tokenStr string) (bool, error) {
 	if err != nil {
 		return false, err
 	}
-
 	if claims.ExpiresAt.Time.Before(time.Now()) {
 		return false, errors.New("token has expired")
 	}
-
 	return true, nil
 }
 
-func Login(w http.ResponseWriter, r *http.Request) {
-	var credentials models.UserCredentials
-	if err := json.NewDecoder(r.Body).Decode(&credentials); err != nil {
-		http.Error(w, "Invalid request", http.StatusBadRequest)
-		return
-	}
-
-	// Fetch user by email using the UserService
-	// Pass the repo as the interface type
-	userRepo := user_repo.NewUserRepository(repo.GetDB())
-    userService := services.NewUserService(userRepo)
-	user, err := userService.GetUserByEmail(credentials.Email)
-	if err != nil {
-		http.Error(w, "Invalid credentials", http.StatusUnauthorized)
-		return
-	}
-
-	// Validate password (bcrypt comparison)
-	err = bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(credentials.Password))
-	if err != nil {
-		http.Error(w, "Invalid credentials", http.StatusUnauthorized)
-		return
-	}
-
-	// Generate JWT token
-	token, err := GenerateToken(int(user.ID))
-	if err != nil {
-		http.Error(w, "Error generating token", http.StatusInternalServerError)
-		return
-	}
-
-	response := map[string]string{
-		"token": token,
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(response)
-}
-
+// checks the token and injects the claims into request context
 func JWTMiddleware(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		authHeader := r.Header.Get("Authorization")
-		if authHeader == "" || !strings.HasPrefix(authHeader, "Bearer ") {
-			http.Error(w, "Missing or invalid Authorization header", http.StatusUnauthorized)
+		tokenStr, err := ExtractTokenFromHeader(r)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusUnauthorized)
 			return
 		}
 
-		tokenStr := strings.TrimPrefix(authHeader, "Bearer ")
 		claims, err := ParseJWT(tokenStr)
 		if err != nil {
 			http.Error(w, "Invalid token: "+err.Error(), http.StatusUnauthorized)
